Polyfill Promise. Fixes #2375
