+++ package.json
@@ -28,8 +28,9 @@
     "lodash": "^4.17.4",
     "messenger": "git+https://github.com/HubSpot/messenger.git",
     "moment": "^2.17.1",
+    "promise": "^7.3.1",
     "sortablejs": "^1.5.1",
-    "sweetalert2": "^6.4.4"
+    "sweetalert2": "^6.6.6"
   },
   "private": true
 }
