+++ public/mix-manifest.json
@@ -1,15 +1,7 @@
 {
+  "/dist/js/app.js": "/dist/js/app.f266b3ff7582a248cf69.js",
   "/mix.js": "/mix.46bdcf356e69eb1117ec.js",
-  "/js/app.js": "/js/app.js",
-  "/js/all.js": "/js/all.96fbf298e0087299ee7bd00e57d3bcc1.js",
-  "/css/dashboard/dashboard.css": "/css/dashboard/dashboard.d62a8b6468ab8c9bca396ab8f6cde506.css",
-  "/css/app.css": "/css/app.feac6a0c1283b11117bc898e9697488a.css",
-  "/mix.js.map": "/mix.46bdcf356e69eb1117ec.js.map",
-  "/css/dashboard/dashboard.css.map": "/css/dashboard/dashboard.d62a8b6468ab8c9bca396ab8f6cde506.css.map",
-  "/css/app.css.map": "/css/app.feac6a0c1283b11117bc898e9697488a.css.map",
-  "/dist/css/dashboard/dashboard.css": "/dist/css/dashboard/dashboard.d62a8b6468ab8c9bca396ab8f6cde506.css",
+  "/dist/css/dashboard/dashboard.css": "/dist/css/dashboard/dashboard.b475b26f586c0747452d537bdf88aec4.css",
   "/dist/css/app.css": "/dist/css/app.feac6a0c1283b11117bc898e9697488a.css",
-  "/dist/css/dashboard/dashboard.css.map": "/dist/css/dashboard/dashboard.d62a8b6468ab8c9bca396ab8f6cde506.css.map",
-  "/dist/css/app.css.map": "/dist/css/app.feac6a0c1283b11117bc898e9697488a.css.map",
-  "/dist/js/all.js": "/dist/js/all.96fbf298e0087299ee7bd00e57d3bcc1.js"
+  "/dist/js/all.js": "/dist/js/all.7086017b336c8e35b8de419383fbdc12.js"
 }
